Add CreateShardingBindingTableRuleStatementUpdater
