+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.RDLUpdater
@@ -16,7 +16,11 @@
 #
 
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterShardingTableRuleStatementUpdater
+
+org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.CreateShardingBindingTableRuleStatementUpdater
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterShardingBindingTableRuleStatementUpdater
+
+
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterShardingBroadcastTableRuleStatementUpdater
 
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.CreateReadwriteSplittingRuleStatementUpdater
